move user_data fixture to conftest.py
